requirements.txt
Django==2.2.6
django-ckeditor>=5.8.0
django-jet==1.0.8
django-js-asset>=1.2.2
django-bootstrap4
django-bootstrap-datepicker-plus==3.0.5
reportlab
django_extensions
django-import-export
